Updated types and added strategy
