+++ src/plugins/piral-oauth2/src/setup.ts
@@ -1,87 +1,6 @@
 import * as ClientOAuth2 from 'client-oauth2';
 import { createOAuth2MemoryPersistence } from './utils';
-import { OAuth2Persistence } from './types';
-
-/**
- * Available configuration options for the OAuth 2 plugin.
- */
-export interface OAuth2Config {
-  /**
-   * The id of the client. Required for the setup of OAuth 2.
-   */
-  clientId: string;
-  /**
-   * The client secret. Only required for the `code` flow.
-   */
-  clientSecret?: string;
-  /**
-   * The Uri pointing to the authorization endpoint of the Identity Provider.
-   */
-  authorizationUri: string;
-  /**
-   * The Uri pointing to the access token endpoint of the Identity Provider.
-   */
-  accessTokenUri?: string;
-  /**
-   * The redirect Uri to use. By default the origin with /auth
-   * is used.
-   */
-  redirectUri?: string;
-  /**
-   * The return path to use in case of the "code" flow. By default the
-   * path will be set to "/".
-   */
-  returnPath?: string;
-  /**
-   * The scopes to be used.
-   */
-  scopes?: Array<string>;
-  /**
-   * The OAuth 2 authorization flow type to be used.
-   */
-  flow?: 'implicit' | 'code';
-  /**
-   * Restricts token sharing such that other integrations, e.g., with
-   * fetch would need to be done manually.
-   * Otherwise, the client is responsive to the `before-fetch` event.
-   */
-  restrict?: boolean;
-  /**
-   * Optional persistence layer for OAuth 2. By default nothing is stored.
-   */
-  persist?: OAuth2Persistence;
-}
-
-export interface OAuth2Request {
-  /**
-   * Sets the headers of the request.
-   * @param headers Headers or a promise to headers.
-   */
-  setHeaders(headers: any): void;
-}
-
-export interface OAuth2Client {
-  /**
-   * Performs a login.
-   */
-  login(): void;
-  /**
-   * Performs a logout.
-   */
-  logout(): void;
-  /**
-   * Gets a token.
-   */
-  token(): Promise<string>;
-  /**
-   * Checks if the user is currently logged in.
-   */
-  account(): boolean;
-  /**
-   * Extends the headers of the provided request.
-   */
-  extendHeaders(req: OAuth2Request): void;
-}
+import { OAuth2Config, OAuth2Client } from './types';
 
 const callbackName = 'oauth2Cb';
 
@@ -98,10 +17,14 @@ export function setupOAuth2Client(config: OAuth2Config): OAuth2Client {
     redirectUri = `${location.origin}/auth`,
     scopes = [],
     flow,
+    headers,
+    query,
+    state,
     restrict = false,
     returnPath = '/',
     persist = createOAuth2MemoryPersistence(),
   } = config;
+
   const client = new ClientOAuth2({
     clientId,
     clientSecret,
@@ -109,7 +32,11 @@ export function setupOAuth2Client(config: OAuth2Config): OAuth2Client {
     authorizationUri,
     accessTokenUri,
     scopes,
+    headers,
+    query,
+    state,
   });
+
   let currentToken: ClientOAuth2.Token;
   let retrieveToken: () => Promise<string>;
   let getLoginUri: () => string;
@@ -192,6 +119,7 @@ export function setupOAuth2Client(config: OAuth2Config): OAuth2Client {
   }
 
   return {
+    _: client,
     login() {
       window.location.href = getLoginUri();
     },
